[design] 채팅 폼 ui
